First try

+++ src/app/list-hopitaux/list-hopitaux.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-
 import { HopitalService } from '../services/hopital.service';
 import { Hopital } from '../models/hopital';
-
+import Swal from 'sweetalert2'
 
 
 @Component({
@@ -27,8 +26,35 @@ export class ListHopitauxComponent implements OnInit {
 
   }
   deletHopital(idHopital : number){
-    this.hopitalService.deleteHopital(idHopital).subscribe(
-      data=> {console.log(data)}
-      )
+        Swal.fire({
+          title: 'Etes-vous-sûr de vouloir supprimer?',
+          text: "Pas de retour en arrière possible",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'le supprimer'
+        }).then((result) => {
+          
+          if (result.value) {
+            this.hopitalService.deleteHopital(idHopital).subscribe(
+              data => {
+
+            Swal.fire(
+              'Supprimer!',
+              'Votre choix a bien été supprimer.',
+              'success'
+            )
+            window.location.href = "http://localhost:4200/hopital"
+          }
+        )}
+        
+        }
+    )
+      }
+      
+
+
+      
   }
-}
+
